requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
defusedxml==0.7.1
Django==6.0.2
djangorestframework==3.16.1
httpie==3.2.4
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.1
psycopg2-binary==2.9.11
Pygments==2.19.2
PySocks==1.7.1
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.2
setuptools==82.0.0
sqlparse==0.5.5
urllib3==2.6.3
django-allauth
PyJWT
cryptography
dj-rest-auth
djangorestframework-simplejwt
celery
redis